Split class
